+++ ethereum/core/src/integration-test/java/org/hyperledger/besu/ethereum/mainnet/precompiles/privacy/PrivacyPrecompiledContractIntegrationTest.java
@@ -52,7 +52,7 @@ import org.hyperledger.besu.evm.tracing.OperationTracer;
 import org.hyperledger.besu.evm.worldstate.WorldUpdater;
 import org.hyperledger.enclave.testutil.EnclaveEncryptorType;
 import org.hyperledger.enclave.testutil.EnclaveKeyConfiguration;
-import org.hyperledger.enclave.testutil.TesseraTestHarness;
+import org.hyperledger.enclave.testutil.EnclaveTestHarness;
 import org.hyperledger.enclave.testutil.TesseraTestHarnessFactory;
 
 import java.nio.file.Files;
@@ -92,7 +92,7 @@ public class PrivacyPrecompiledContractIntegrationTest {
   private static Enclave enclave;
   private static MessageFrame messageFrame;
 
-  private static TesseraTestHarness testHarness;
+  private static EnclaveTestHarness testHarness;
   private static WorldStateArchive worldStateArchive;
   private static PrivateStateStorage privateStateStorage;
   private static final Vertx vertx = Vertx.vertx();
@@ -128,8 +128,7 @@ public class PrivacyPrecompiledContractIntegrationTest {
             new EnclaveKeyConfiguration(
                 new String[] {"enclave_key_0.pub"},
                 new String[] {"enclave_key_1.key"},
-                EnclaveEncryptorType.NOOP),
-            Optional.empty());
+                EnclaveEncryptorType.NOOP));
 
     testHarness.start();